move project check in buttons into own component
