+++ src/api/database.json
@@ -136,7 +136,7 @@
       "name": "Office Renovation",
       "startdate": "2025-03-01",
       "enddate": "2025-06-01",
-      "ageSinceTouch": 0,
+      "ageSinceTouch": 3,
       "description": ""
     },
     {
